2020.03.15 AM00
[1/3] fix a bug in the sample source code written in Fortran 2008
[2/3] add a sample source code written in Fortran 2008
[3/3] update .gitignore

+++ fortran/mini_parts/isMultiple.f90
@@ -259,7 +259,7 @@ module mod_isMultiple
     pure function isOdd_mod_64 (i) result(stat)
 
         ! arguments for this <function>
-        integer(INT16), intent(in) :: i
+        integer(INT64), intent(in) :: i
 
         ! return value of this <function>
         logical :: stat
